requirements.txt
zai
PyPDF2
python-dotenv